# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/segment/": {
            "post": {
                "description": "Метод создания сегмента. Принимает slug (название) сегмента.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "segment"
                ],
                "summary": "Метод создания сегмента.",
                "parameters": [
                    {
                        "example": "{\"slug\": \"AVITO_DISCOUNT_30\"}",
                        "description": "Название сегмента",
                        "name": "slug",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/segment/{slug}": {
            "delete": {
                "description": "Метод удаления сегмента. Принимает slug (название) сегмента.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "segment"
                ],
                "summary": "Метод удаления сегмента.",
                "parameters": [
                    {
                        "type": "string",
                        "example": "AVITO_DISCOUNT_30",
                        "description": "Название сегмента",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/{user_id}/segments": {
            "get": {
                "description": "Метод получения активных сегментов пользователя. Принимает на вход id пользователя.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Метод получения активных сегментов пользователя.",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 1000,
                        "description": "ID пользователя",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Метод добавления/удаления пользователя в сегмент. Принимает список slug (названий) сегментов которые нужно добавить пользователю, список slug (названий) сегментов которые нужно удалить у пользователя, id пользователя.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Метод добавления/удаления пользователя в сегмент.",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 1000,
                        "description": "ID пользователя",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Список сегментов",
                        "name": "segments",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/structs.SegmentAction"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/{user_id}/segments/history/{period}": {
            "get": {
                "description": "Метод получения истории изменения сегментов пользователя. Принимает на вход id пользователя и период год-месяц.",
                "produces": [
                    "text/csv"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Метод получения истории изменения сегментов пользователя.",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 1000,
                        "description": "ID пользователя",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "2023-08",
                        "description": "период год-месяц",
                        "name": "period",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "structs.SegmentAction": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        }
    }
}